chore: only add consumer_version_id index when not mysql

+++ db/migrations/20230131_add_cons_ver_id_ndx_to_latest_pp_id_for_cons_ver.rb
@@ -4,14 +4,18 @@ include PactBroker::MigrationHelper
 
 Sequel.migration do
   up do
-    alter_table(:latest_pact_publication_ids_for_consumer_versions) do
-      add_index([:consumer_version_id], name: "latest_pp_ids_for_cons_ver_con_ver_id_ndx")
+    if !mysql?
+      alter_table(:latest_pact_publication_ids_for_consumer_versions) do
+        add_index([:consumer_version_id], name: "latest_pp_ids_for_cons_ver_con_ver_id_ndx")
+      end
     end
   end
 
   down do
-    alter_table(:latest_pact_publication_ids_for_consumer_versions) do
-      drop_index([:consumer_version_id], name: "latest_pp_ids_for_cons_ver_con_ver_id_ndx")
+    if !mysql?
+      alter_table(:latest_pact_publication_ids_for_consumer_versions) do
+        drop_index([:consumer_version_id], name: "latest_pp_ids_for_cons_ver_con_ver_id_ndx")
+      end
     end
   end
 end
